move most of query params code to one place
