+++ designer/client/src/windowManager/useWindows.ts
@@ -1,27 +1,11 @@
 import { useWindowManager, WindowId, WindowType } from "@touk/window-manager";
-import { defaults, uniq, without } from "lodash";
-import * as queryString from "query-string";
+import { defaults } from "lodash";
 import { useCallback, useMemo } from "react";
-import { ensureArray } from "../common/arrayUtils";
 import { useUserSettings } from "../common/userSettings";
-import { defaultArrayFormat } from "../common/VisualizationUrl";
 import { ConfirmDialogData } from "../components/modals/GenericConfirmDialog";
 import { NodeType, Process } from "../types";
 import { WindowKind } from "./WindowKind";
 import { InfoDialogData } from "../components/modals/GenericInfoDialog";
-
-export function parseWindowsQueryParams<P extends Record<string, string | string[]>>(append: P, remove?: P): Record<string, string[]> {
-    const query = queryString.parse(window.location.search, { arrayFormat: defaultArrayFormat });
-    const keys = uniq(Object.keys({ ...append, ...remove }));
-    return Object.fromEntries(
-        keys.map((key) => {
-            const current = ensureArray(query[key]).map(decodeURIComponent);
-            const withAdded = uniq(current.concat(append?.[key]));
-            const cleaned = without(withAdded, ...ensureArray(remove?.[key])).filter(Boolean);
-            return [key, cleaned];
-        }),
-    );
-}
 
 export function useWindows(parent?: WindowId) {
     const { open: _open, closeAll } = useWindowManager(parent);
